moved traceprovider and span to seperate files
